refactor(angular): extract shared ForesightBase for directive and component
ForesightDirective and ForesightComponent duplicated the registration
signal, state computed, apply/teardown logic, and ngOnDestroy. Both are
now thin subclasses that only declare inputs and map them to
resolveOptions. Also removes the per-instance fallbackState signal that
was never written, in favor of the shared UNREGISTERED_STATE constant.
